[merge] company : 기업 상세보기 페이지 완료, 기업 인가 관련 기능 role 나누기 완료
* company : 기업회원 삭제 시에 해당 기업이 작성한 공고들도 전부 삭제

* company : 기업 상세보기 - 로그인 시에 구독확인 출력 완료,README.md 수정

* company : 기업 상세보기 - 로그인에role 역할 나누기

+++ src/main/java/site/metacoding/humancloud/service/CompanyService.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.regex.Pattern;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,7 +22,6 @@ import site.metacoding.humancloud.domain.recruit.Recruit;
 import site.metacoding.humancloud.domain.recruit.RecruitDao;
 import site.metacoding.humancloud.domain.resume.Resume;
 import site.metacoding.humancloud.domain.resume.ResumeDao;
-import site.metacoding.humancloud.domain.subscribe.Subscribe;
 import site.metacoding.humancloud.domain.subscribe.SubscribeDao;
 import site.metacoding.humancloud.domain.user.UserDao;
 import site.metacoding.humancloud.dto.auth.UserFindByAllUsernameDto;
@@ -31,6 +31,7 @@ import site.metacoding.humancloud.dto.company.CompanyRespDto.CompanyDetailRespDt
 import site.metacoding.humancloud.dto.company.CompanyRespDto.CompanyFindById;
 import site.metacoding.humancloud.dto.company.CompanyRespDto.CompanyUpdateRespDto;
 import site.metacoding.humancloud.dto.dummy.response.page.PagingDto;
+import site.metacoding.humancloud.dto.recruit.RecruitRespDto.RecruitListByCompanyIdRespDto;
 import site.metacoding.humancloud.util.SHA256;
 
 @Slf4j
@@ -93,11 +94,11 @@ public class CompanyService {
 	@Transactional(readOnly = true)
 	public CompanyDetailRespDto 기업정보상세보기(Integer userId, Integer companyId) {
 		Optional<CompanyFindById> companyOP = companyDao.findById(companyId);
-		if (subscribeDao.findById(userId, companyId) == null) {
-			return false;
+		boolean isSub = false;
+		if (subscribeDao.findById(userId, companyId) != null) {
+			isSub = true;
 		}
-		return true;
-		CompanyDetailRespDto companyPS = new CompanyDetailRespDto(companyOP.get(), true);
+		CompanyDetailRespDto companyPS = new CompanyDetailRespDto(companyOP.get(), isSub);
 
 		// 전화번호 포매팅
 		String fomat = "(\\d{2,3})(\\d{3,4})(\\d{4})";
@@ -178,19 +179,18 @@ public class CompanyService {
 		companyOP.orElseThrow(() -> new RuntimeException("존재하지 않는 회원입니다."));
 
 		// 해당 Company의 채용공고 삭제
-		List<Recruit> recruits = recruitDao.findByCompanyId(id);
-		if (recruits != null) {
-			recruitDao.deleteById(id);
+		Optional<List<RecruitListByCompanyIdRespDto>> recruitListOP = recruitDao.findByCompanyId(id);
+		if (recruitListOP.isPresent()) {
+			companyDao.deleteById(id);
 		}
 
-		companyDao.deleteById(id);
 	}
 
-	public List<Recruit> 채용공고리스트불러오기(Integer id) {
-		for (int i = 0; i < recruitDao.findByCompanyId(id).size(); i++) {
-			System.out.println(recruitDao.findByCompanyId(id).get(i).getRecruitTitle());
+	public List<RecruitListByCompanyIdRespDto> 채용공고리스트불러오기(Integer id) {
+		for (int i = 0; i < recruitDao.findByCompanyId(id).get().size(); i++) {
+			System.out.println(recruitDao.findByCompanyId(id).get().get(i).getRecruitTitle());
 		}
-		return recruitDao.findByCompanyId(id);
+		return recruitDao.findByCompanyId(id).get();
 	}
 
 	public List<Resume> 지원목록보기(Integer companyId) {
